# swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Huynh Land Firebase API
  version: 1.0.0
paths:
  /api/history:
    get:
      tags: [History]
      summary: Hiển thị lịch sử hình thành theo ngôn ngữ
      parameters:
        - in: query
          name: lang
          required: true
          schema:
            type: string
            enum: [vi, en, ko]
          description: Ngôn ngữ (vi, en, ko)
      responses:
        "200":
          description: Lấy lịch sử thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  history:
                    type: object
        "400":
          description: Ngôn ngữ không hợp lệ
        "404":
          description: Không có lịch sử
    post:
      tags: [History]
      summary: Tạo lịch sử hình thành theo ngôn ngữ
      parameters:
        - in: query
          name: lang
          required: true
          schema:
            type: string
            enum: [vi, en, ko]
          description: Ngôn ngữ (vi, en, ko)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                content:
                  type: string
      responses:
        "201":
          description: Lịch sử đã được tạo
        "400":
          description: Lịch sử đã tồn tại hoặc ngôn ngữ không hợp lệ
    put:
      tags: [History]
      summary: Sửa lịch sử hình thành theo ngôn ngữ
      parameters:
        - in: query
          name: lang
          required: true
          schema:
            type: string
            enum: [vi, en, ko]
          description: Ngôn ngữ (vi, en, ko)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                content:
                  type: string
      responses:
        "200":
          description: Lịch sử đã được cập nhật
        "400":
          description: Ngôn ngữ không hợp lệ
        "404":
          description: Không có lịch sử để cập nhật
    delete:
      tags: [History]
      summary: Xóa lịch sử hình thành theo ngôn ngữ
      parameters:
        - in: query
          name: lang
          required: true
          schema:
            type: string
            enum: [vi, en, ko]
          description: Ngôn ngữ (vi, en, ko)
      responses:
        "200":
          description: Lịch sử đã được xóa
        "400":
          description: Ngôn ngữ không hợp lệ
        "404":
          description: Không có lịch sử để xóa
...
  /api/contact:
    get:
      tags: [Contact]
      summary: Lấy thông tin liên hệ
      responses:
        "200":
          description: Lấy thông tin liên hệ thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  contact:
                    type: object
                    properties:
                      hotline:
                        type: string
                      email:
                        type: string
                      workingHours:
                        type: string
        "404":
          description: Không có thông tin liên hệ
    post:
      tags: [Contact]
      summary: Tạo thông tin liên hệ
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                hotline:
                  type: string
                email:
                  type: string
                workingHours:
                  type: string
      responses:
        "201":
          description: Đã tạo thông tin liên hệ
        "400":
          description: Thông tin liên hệ đã tồn tại hoặc thiếu trường
    put:
      tags: [Contact]
      summary: Sửa thông tin liên hệ
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                hotline:
                  type: string
                email:
                  type: string
                workingHours:
                  type: string
      responses:
        "200":
          description: Đã cập nhật thông tin liên hệ
        "404":
          description: Không có thông tin liên hệ để cập nhật
    delete:
      tags: [Contact]
      summary: Xóa thông tin liên hệ
      responses:
        "200":
          description: Đã xóa thông tin liên hệ
        "404":
          description: Không có thông tin liên hệ để xóa
  /api/office:
    get:
      tags: [Office]
      summary: Hiển thị địa chỉ văn phòng theo ngôn ngữ
      parameters:
        - in: query
          name: lang
          required: true
          schema:
            type: string
            enum: [vi, en, ko]
          description: Ngôn ngữ (vi, en, ko)
      responses:
        "200":
          description: Lấy địa chỉ văn phòng thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  office:
                    type: object
                    properties:
                      name:
                        type: string
                      address:
                        type: string
                      phone:
                        type: string
                      gmail:
                        type: string
        "400":
          description: Ngôn ngữ không hợp lệ
        "404":
          description: Không có địa chỉ văn phòng
    post:
      tags: [Office]
      summary: Tạo địa chỉ văn phòng theo ngôn ngữ
      parameters:
        - in: query
          name: lang
          required: true
          schema:
            type: string
            enum: [vi, en, ko]
          description: Ngôn ngữ (vi, en, ko)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                address:
                  type: string
                phone:
                  type: string
                gmail:
                  type: string
      responses:
        "201":
          description: Đã tạo địa chỉ văn phòng
        "400":
          description: Địa chỉ văn phòng đã tồn tại hoặc thiếu trường
    put:
      tags: [Office]
      summary: Sửa địa chỉ văn phòng theo ngôn ngữ
      parameters:
        - in: query
          name: lang
          required: true
          schema:
            type: string
            enum: [vi, en, ko]
          description: Ngôn ngữ (vi, en, ko)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                address:
                  type: string
                phone:
                  type: string
                gmail:
                  type: string
      responses:
        "200":
          description: Đã cập nhật địa chỉ văn phòng
        "400":
          description: Ngôn ngữ không hợp lệ
        "404":
          description: Không có địa chỉ văn phòng để cập nhật
    delete:
      tags: [Office]
      summary: Xóa địa chỉ văn phòng theo ngôn ngữ
      parameters:
        - in: query
          name: lang
          required: true
          schema:
            type: string
            enum: [vi, en, ko]
          description: Ngôn ngữ (vi, en, ko)
      responses:
        "200":
          description: Đã xóa địa chỉ văn phòng
        "400":
          description: Ngôn ngữ không hợp lệ
        "404":
          description: Không có địa chỉ văn phòng để xóa
...
  /api/mainoffice-map:
    get:
      tags: [MainOfficeMap]
      summary: Lấy thông tin bản đồ văn phòng chính
      responses:
        "200":
          description: Lấy thông tin bản đồ văn phòng chính thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  mainOfficeMap:
                    type: object
                    properties:
                      address:
                        type: string
                      iframe:
                        type: string
        "404":
          description: Không có thông tin bản đồ văn phòng chính
    post:
      tags: [MainOfficeMap]
      summary: Tạo thông tin bản đồ văn phòng chính
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                address:
                  type: string
                iframe:
                  type: string
      responses:
        "201":
          description: Đã tạo thông tin bản đồ văn phòng chính
        "400":
          description: Đã tồn tại hoặc thiếu trường
    put:
      tags: [MainOfficeMap]
      summary: Sửa thông tin bản đồ văn phòng chính
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                address:
                  type: string
                iframe:
                  type: string
      responses:
        "200":
          description: Đã cập nhật thông tin bản đồ văn phòng chính
        "404":
          description: Không có thông tin để cập nhật
    delete:
      tags: [MainOfficeMap]
      summary: Xóa thông tin bản đồ văn phòng chính
      responses:
        "200":
          description: Đã xóa thông tin bản đồ văn phòng chính
        "404":
          description: Không có thông tin để xóa
...